Add files via upload
Created a concise, quick, one-page interactive dashboard to analyze the sale of beverages based on Category, Class and Province spanning a year.
EWS is distributing wines and spirits to the majority areas of Canada, the United States (Nevada), a few parts of South America, South Africa, and Australia  
Demonstrated a geographical representation of products being distributed throughout the vendors across the world 
Identified the top 5 global sales and represented them in a line graph based on Class, Category and Province that helps EWS choose products to sell to the retailers who would generate high sales in future 
Scrutinized and detached the Product’s unique number from the Product name (For ex. Separated 699860 from Unauthorized Supplier - 12x1000) into an individual column to link the product’s unique no. with the same in other excel documents to distinguish for easier access, analysis, and forecasting 
Classified the Class, Categories, and Unique Product number based on period with a horizontal bar chart and numbers for achieving clarity, transparency, and ease of understanding. Highlighted the selling rate in different colours (0-33%-> Red, 34%-66%->Yellow, 67%-100%-> Green) 
Recognized that most of the products have ‘0’ as their Product number, which might cause duplication and complexity, therefore can give an incorrect prediction 
There are around 2628069 Categories without Class (Parent). For ex. Cider, 7% Sparkling, Alcohol, Mead, Ice Wine, and Domestic Beer do not fall under any Class 
Created Slicers with a drop-down to filter products based on Class, Category and Province for ease of access